+++ src/main/java/com/devwarriors/mapfood/MapFoodApplication.java
@@ -1,0 +1,13 @@
+package com.devwarriors.mapfood;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class MapFoodApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(MapFoodApplication.class, args);
+	}
+
+}
