+++ src/main/java/com/devwarriors/mapfood/dto/ItemPedidoDto.java
@@ -1,0 +1,15 @@
+package com.devwarriors.mapfood.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@AllArgsConstructor
+@Getter
+public class ItemPedidoDto {
+
+	private String produtoId;
+	private Integer quantidade;
+	private String observacao;
+}
