+++ src/main/java/com/devwarriors/mapfood/dto/PedidoDto.java
@@ -1,0 +1,4 @@
+package com.devwarriors.mapfood.dto;
+
+public interface PedidoDto {
+}
