+++ src/main/java/com/devwarriors/mapfood/model/PedidoStatus.java
@@ -1,0 +1,5 @@
+package com.devwarriors.mapfood.model;
+
+public enum PedidoStatus {
+	EM_PREPARO, EM_TRANSITO, ENTREGUE
+}
