+++ src/main/java/com/devwarriors/mapfood/repository/EntregadorRepository.java
@@ -1,0 +1,8 @@
+package com.devwarriors.mapfood.repository;
+
+import com.devwarriors.mapfood.model.Entregador;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+public interface EntregadorRepository extends MongoRepository<Entregador, String>{
+
+}
