+++ src/main/java/com/devwarriors/mapfood/repository/PedidoRepository.java
@@ -1,0 +1,7 @@
+package com.devwarriors.mapfood.repository;
+
+import com.devwarriors.mapfood.model.Pedido;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+public interface PedidoRepository extends MongoRepository<Pedido, String> {
+}
