--- src/main/java/com/devwarrios/mapfood/MapFoodApplication.java
@@ -1,13 +1,0 @@
-package com.devwarrios.mapfood;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MapFoodApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(MapFoodApplication.class, args);
-	}
-
-}
