--- src/main/java/com/devwarrios/mapfood/dto/PedidoDto.java
@@ -1,4 +1,0 @@
-package com.devwarrios.mapfood.dto;
-
-public interface PedidoDto {
-}
