--- src/main/java/com/devwarrios/mapfood/dto/PedidoRequestDto.java
@@ -1,17 +1,0 @@
-package com.devwarrios.mapfood.dto;
-
-import java.util.List;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class PedidoRequestDto implements PedidoDto {
-
-	private Integer clienteId;
-	private String estabelecimentoId;
-	private List<ItemPedidoDto> itens;
-}
