--- src/main/java/com/devwarrios/mapfood/dto/PedidoResponseDto.java
@@ -1,19 +1,0 @@
-package com.devwarrios.mapfood.dto;
-
-import java.time.LocalDate;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class PedidoResponseDto {
-
-	private String pedidoId;
-	private String entregador;
-	private double distancia;
-	private double eta;
-	private LocalDate timestamp;
-}
