--- src/main/java/com/devwarrios/mapfood/dto/factory/PedidoFactory.java
@@ -1,20 +1,0 @@
-package com.devwarrios.mapfood.dto.factory;
-
-import java.time.LocalDate;
-import java.util.List;
-import java.util.Optional;
-
-import com.devwarrios.mapfood.model.Cliente;
-import com.devwarrios.mapfood.model.Estabelecimento;
-import com.devwarrios.mapfood.model.ItemPedido;
-import com.devwarrios.mapfood.model.Pedido;
-import com.devwarrios.mapfood.model.PedidoStatus;
-
-public class PedidoFactory {
-
-	public static Pedido criaNovoPedido(Optional<Cliente> cliente, Estabelecimento estabelecimento, List<ItemPedido> itens) {
-		Pedido pedido = new Pedido(cliente, estabelecimento, itens, LocalDate.now(), PedidoStatus.EM_PREPARO);
-
-		return pedido;
-	}
-}
