--- src/main/java/com/devwarrios/mapfood/model/PedidoStatus.java
@@ -1,5 +1,0 @@
-package com.devwarrios.mapfood.model;
-
-public enum PedidoStatus {
-	EM_PREPARO, EM_TRANSITO, ENTREGUE
-}
