--- src/main/java/com/devwarrios/mapfood/repository/EntregadorRepository.java
@@ -1,9 +1,0 @@
-package com.devwarrios.mapfood.repository;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import com.devwarrios.mapfood.model.Entregador;
-
-public interface EntregadorRepository extends MongoRepository<Entregador, String>{
-
-}
