--- src/main/java/com/devwarrios/mapfood/repository/PedidoRepository.java
@@ -1,8 +1,0 @@
-package com.devwarrios.mapfood.repository;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-import com.devwarrios.mapfood.model.Pedido;
-
-public interface PedidoRepository extends MongoRepository<Pedido, String> {
-}
