--- src/main/java/com/devwarrios/mapfood/service/ClienteService.java
@@ -1,25 +1,0 @@
-package com.devwarrios.mapfood.service;
-
-import java.util.List;
-import java.util.Optional;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import com.devwarrios.mapfood.model.Cliente;
-import com.devwarrios.mapfood.repository.ClienteRepository;
-
-@Service
-public class ClienteService {
-
-	@Autowired
-	private ClienteRepository clienteRepository;
-
-	public List<Cliente> getClientes() { return clienteRepository.findAll();	}
-
-	public Cliente getClienteById(Integer id) { return clienteRepository.findByClienteId(id);	}
-
-	public Cliente criaCliente(Cliente cliente) {
-		return clienteRepository.save(cliente);
-	}
-}
