--- src/test/java/com/devwarrios/mapfood/MapFoodApplicationTests.java
@@ -1,16 +1,0 @@
-package com.devwarrios.mapfood;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class MapFoodApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
